# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MeshDepot API
  version: 1.0.0
  description: >
    REST API for MeshDepot - a self-hosted 3D design library manager.
    All endpoints (except `/health`, `/auth/login`, `/auth/totp/verify`,
    and `/users/{id}/avatar` GET) require an authenticated session cookie.
    Admin endpoints additionally require the session user to have admin privileges.
servers:
  - url: /api
    description: Relative path (same origin as frontend)
...
components:
  schemas:
...
    Design:
      type: object
      properties:
        id:              { type: string, description: Public id - the design is addressed by this everywhere }
        name:            { type: string }
        description:     { type: string, nullable: true }
        source_url:      { type: string, nullable: true }
        source_platform: { type: string, nullable: true, example: thingiverse }
        source_id:       { type: string, nullable: true }
        category:        { type: string, nullable: true }
        license:         { type: string, nullable: true }
        author:          { type: string, nullable: true }
        rating:          { type: integer, nullable: true, minimum: 1, maximum: 5 }
        print_time_minutes: { type: integer, nullable: true }
        notes:           { type: string, nullable: true }
        cover_path:      { type: string, nullable: true }
        created_at:      { type: string, format: date-time }
        updated_at:      { type: string, format: date-time }
    DesignDetail:
      allOf:
        - $ref: '#/components/schemas/Design'
        - type: object
          properties:
            tags:    { type: array, items: { $ref: '#/components/schemas/Tag' } }
            images:  { type: array, items: { $ref: '#/components/schemas/DesignImage' } }
            shares:  { type: array, items: { $ref: '#/components/schemas/Share' } }
    Tag:
      type: object
      properties:
        id:          { type: integer }
        name:        { type: string }
        color:       { type: string, nullable: true, example: '#e74c3c' }
        usage_count: { type: integer }
...
    DesignImage:
      type: object
      properties:
        id:       { type: integer }
        path:     { type: string }
        sort_order: { type: integer }
...
    Share:
      type: object
      properties:
        id:                  { type: integer }
        shared_with_name:    { type: string }
        shared_with_email:   { type: string, format: email }
        created_at:          { type: string, format: date-time }
...
    Error:
      type: object
      properties:
        error: { type: string }
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: PHPSESSID
      description: PHP session cookie set on login
security:
  - sessionCookie: []
paths:
...
  /designs:
    get:
      tags: [Designs]
      summary: List designs
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: source_platform, in: query, schema: { type: string } }
        - { name: tag_ids, in: query, description: Comma-separated tag IDs, schema: { type: string } }
        - { name: shared_only, in: query, schema: { type: boolean } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: per_page, in: query, schema: { type: integer, default: 24 } }
      responses:
        '200':
          description: Paginated design list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Design' } }
                  total: { type: integer }
    post:
      tags: [Designs]
      summary: Create design
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:            { type: string }
                description:     { type: string }
                source_url:      { type: string }
                source_platform: { type: string }
                source_id:       { type: string }
                category:        { type: string }
                license:         { type: string }
                author:          { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Design' }
...
  /designs/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string }, description: The design's public id }
    get:
      tags: [Designs]
      summary: Get design detail
      responses:
        '200':
          description: Design with tags, images and shares
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DesignDetail' }
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Designs]
      summary: Update design metadata
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:               { type: string }
                description:        { type: string }
                source_url:         { type: string }
                source_platform:    { type: string }
                source_id:          { type: string }
                category:           { type: string }
                license:            { type: string }
                author:             { type: string }
                rating:             { type: integer, minimum: 1, maximum: 5 }
                print_time_minutes: { type: integer }
                notes:              { type: string }
      responses:
        '200':
          description: Updated
    delete:
      tags: [Designs]
      summary: Delete design and all associated files
      responses:
        '200':
          description: Deleted
...
  /designs/{id}/images:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string }, description: The design's public id }
    post:
      tags: [Design Images]
      summary: Upload gallery images
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  description: One or more image files (JPEG/PNG/WebP/GIF)
                  oneOf:
                    - { type: string, format: binary }
                    - { type: array, items: { type: string, format: binary } }
      responses:
        '200':
          description: Number of images uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  uploaded: { type: integer }
  /designs/{id}/images/{imageId}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string }, description: The design's public id }
      - { name: imageId, in: path, required: true, schema: { type: integer } }
    delete:
      tags: [Design Images]
      summary: Delete gallery image
      responses:
        '200':
          description: Deleted
...
  /collections/{id}/designs:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    get:
      tags: [Collections]
      summary: List designs in collection
      responses:
        '200':
          description: Design list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Design' }
    post:
      tags: [Collections]
      summary: Add designs to collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [design_ids]
              properties:
                design_ids: { type: array, items: { type: integer } }
      responses:
        '200':
          description: Added
...
  /collections/{id}/designs/{designId}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
      - { name: designId, in: path, required: true, schema: { type: string }, description: The design's public id }
    delete:
      tags: [Collections]
      summary: Remove design from collection
      responses:
        '200':
          description: Removed

  # ── Downloads ────────────────────────────────────────────────────────────────
...